# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "business-project API",
    "description": "Контракт между бэкендом и клиентами",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "account"
        ],
        "operationId": "update_me",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMe"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "Тело ответа с ошибкой — единственный формат на весь API.\n`fields` появляется только у ошибок про конкретные поля формы.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "машиночитаемый код (он же ключ локализации)"
          },
          "fields": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "message": {
            "type": "string",
            "description": "текст на языке запроса — можно показывать пользователю как есть"
          }
        }
      },
      "FieldError": {
        "type": "object",
        "required": [
          "field",
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "field": {
            "type": "string",
            "description": "имя поля в теле запроса, например \"password\""
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "user",
          "admin"
        ]
      },
...
      "UpdateMe": {
        "type": "object",
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "как обращаться к пользователю; пропущено — не меняем",
            "maxLength": 100
          },
          "locale": {
            "type": [
              "string",
              "null"
            ],
            "description": "язык интерфейса и писем: \"ru\" или \"en\""
          }
        }
      },
      "User": {
        "type": "object",
        "required": [
          "id",
          "email",
          "role",
          "totp_enabled",
          "email_verified"
        ],
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "как обращаться к пользователю; не задано — показываем почту"
          },
          "email": {
            "type": "string"
          },
          "email_verified": {
            "type": "boolean",
            "description": "подтверждена ли почта (по ссылке из письма)"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "locale": {
            "type": [
              "string",
              "null"
            ],
            "description": "язык интерфейса и писем: код из i18n::ALL_LANGS"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "totp_enabled": {
            "type": "boolean",
            "description": "включён ли второй фактор"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}